processus_etl_sante/requirements.txt
agate==1.6.1
alembic==1.7.4
annotated-types==0.6.0
anyio==3.3.3
apache-airflow==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-google==6.0.0
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Automat==0.8.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
blinker==1.4
cachetools==4.2.2
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
clickclick==20.10.2
cloud-init==23.1.1
colorama==0.4.4
colorlog==4.8.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
croniter==1.0.15
cryptography==3.4.8
dbt-core==0.21.1
dbt-extractor==0.4.0
dbt-semantic-interfaces==0.4.4
dbus-python==1.2.16
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
eventlet==0.32.0
filelock==3.3.0
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.8.0
google-ads==14.0.0
google-api-core==1.31.3
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==0.1.5
google-cloud-audit-log==0.1.1
google-cloud-automl==2.4.2
google-cloud-bigquery==2.28.1
google-cloud-bigquery-datatransfer==3.3.4
google-cloud-bigquery-storage==2.9.1
google-cloud-bigtable==1.7.0
google-cloud-build==3.5.2
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.4.3
google-cloud-dataproc==2.5.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.8.0
google-cloud-language==1.3.0
google-cloud-logging==2.6.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.6.0
google-cloud-os-login==2.3.2
google-cloud-pubsub==2.8.0
google-cloud-redis==2.2.4
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.42.3
google-cloud-tasks==2.6.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.2.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
graphviz==0.17
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
grpcio-gcp==0.2.2
gunicorn==20.1.0
h11==0.12.0
hologram==0.0.14
httpcore==0.13.7
httplib2==0.19.1
httpx==0.19.0
hyperlink==19.0.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==16.10.1
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
json-rpc==1.15.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
leather==0.4.0
libcst==0.3.21
lockfile==0.12.2
Logbook==1.5.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
mashumaro==2.5
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==2.8.8
nox==2020.12.31
numpy==1.20.3
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
pandas==1.3.3
pandas-gbq==0.14.1
parsedatetime==2.6
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.6.0
platformdirs==2.4.0
prison==0.2.1
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
py==1.10.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.6.4
pydantic-core==2.16.3
pydata-google-auth==1.2.0
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==10.12.0
rsa==4.7.2
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.2.2
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
sos==4.4
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.4
ssh-import-id==5.10
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
Twisted==18.9.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.7
virtualenv==20.8.1
wadllib==1.3.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
